+++ data-transfer/src/test/java/it/eng/datatransfer/rest/protocol/ConsumerDataTransferCallbackControllerTest.java
@@ -13,9 +13,13 @@ import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+
+import com.fasterxml.jackson.databind.JsonNode;
 
 import it.eng.datatransfer.exceptions.TransferProcessNotFoundException;
 import it.eng.datatransfer.model.Serializer;
+import it.eng.datatransfer.model.TransferCompletionMessage;
 import it.eng.datatransfer.model.TransferStartMessage;
 import it.eng.datatransfer.service.DataTransferService;
 import it.eng.datatransfer.util.MockObjectUtil;
@@ -61,10 +65,12 @@ public class ConsumerDataTransferCallbackControllerTest {
 	@Test
 	@DisplayName("Complete TransferProcess")
 	public void completeDataTransfer() {
-		assertEquals(HttpStatus.NOT_IMPLEMENTED, 
-				controller.completeDataTransfer(MockObjectUtil.CONSUMER_PID,
-						Serializer.serializeProtocolJsonNode(MockObjectUtil.TRANSFER_COMPLETION_MESSAGE)).getStatusCode());
-	}
+		when(dataTransferService.completeDataTransfer(any(TransferCompletionMessage.class), any(String.class), isNull()))
+		.thenReturn(MockObjectUtil.TRANSFER_PROCESS_COMPLETED);
+		ResponseEntity<JsonNode> response = controller.completeDataTransfer(MockObjectUtil.CONSUMER_PID,
+				Serializer.serializeProtocolJsonNode(MockObjectUtil.TRANSFER_COMPLETION_MESSAGE));
+		assertEquals(HttpStatus.OK, response.getStatusCode());	
+		}
 	
 	@Test
 	@DisplayName("Complete TransferProcess - invalid request body")
@@ -72,6 +78,16 @@ public class ConsumerDataTransferCallbackControllerTest {
 		assertThrows(ValidationException.class, () ->
 			controller.completeDataTransfer(MockObjectUtil.CONSUMER_PID, 
 					Serializer.serializeProtocolJsonNode(MockObjectUtil.TRANSFER_START_MESSAGE)));
+	}
+	
+	@Test
+	@DisplayName("Complete TransferProcess - error service")
+	public void completeDataTransfer_errorService() {
+		when(dataTransferService.completeDataTransfer(any(TransferCompletionMessage.class), any(String.class), isNull()))
+			.thenThrow(TransferProcessNotFoundException.class);
+		assertThrows(TransferProcessNotFoundException.class, 
+				() -> controller.completeDataTransfer(MockObjectUtil.PROVIDER_PID,
+				Serializer.serializeProtocolJsonNode(MockObjectUtil.TRANSFER_COMPLETION_MESSAGE)));
 	}
 	
 	@Test